01_LLM_Agents/multimodal-agent-agrobot/requirements.txt
# AgroBot + AI Multimodal Agent — Dependențe Python
# pip install -r requirements.txt

# ML / Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Vizualizare (opțional, pentru grafice)
matplotlib>=3.7.0

# Interfață web
gradio>=4.0.0

# Procesare imagini
Pillow>=10.0.0

# Voce / Audio
SpeechRecognition>=3.10.0
gtts>=2.3.0

# PyAudio (transcriere microfon — instalare separată)
# pip install pyaudio  (macOS: brew install portaudio first)
